+++ packages/ApiTests/src/dapi/functions.ts
@@ -347,7 +347,7 @@ category('Dapi: functions calls', async () => {
     await grok.dapi.functions.calls.delete(funcCall);
     expect(await grok.dapi.functions.calls.find(funcCall.id) === undefined, true, 'funcCall was not deleted');
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});
 
 category('Dapi: functions', async () => {
   test('Load package function with package', async () => {
@@ -393,4 +393,4 @@ category('Dapi: functions', async () => {
     await grok.dapi.functions.calls.allPackageVersions()
       .include('session.user,func.package, inputs, outputs').find(fc.id);
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});